+++ packages/flutter_riverpod/test/consumer_test.dart
@@ -41,10 +41,10 @@ void main() {
 
     container.read(provider0.state);
     container.read(provider1.state);
-    final familyState0 = container.debugProviderStates.firstWhere((p) {
+    final familyState0 = container.debugProviderElements.firstWhere((p) {
       return p.provider == provider0.state;
     });
-    final familyState1 = container.debugProviderStates.firstWhere((p) {
+    final familyState1 = container.debugProviderElements.firstWhere((p) {
       return p.provider == provider1.state;
     });
 
@@ -114,10 +114,10 @@ void main() {
 
     container.read(provider0.state);
     container.read(provider1.state);
-    final familyState0 = container.debugProviderStates.firstWhere((p) {
+    final familyState0 = container.debugProviderElements.firstWhere((p) {
       return p.provider == provider0.state;
     });
-    final familyState1 = container.debugProviderStates.firstWhere((p) {
+    final familyState1 = container.debugProviderElements.firstWhere((p) {
       return p.provider == provider1.state;
     });
 
@@ -311,7 +311,7 @@ void main() {
         .firstState<ProviderScopeState>(find.byKey(firstOwnerKey))
         .container;
 
-    final state1 = owner1.debugProviderStates
+    final state1 = owner1.debugProviderElements
         .firstWhere((s) => s.provider == provider.state);
 
     expect(state1.hasListeners, true);
@@ -339,7 +339,7 @@ void main() {
         .firstState<ProviderScopeState>(find.byKey(secondOwnerKey))
         .container;
 
-    final state2 = container2.debugProviderStates
+    final state2 = container2.debugProviderElements
         .firstWhere((s) => s.provider is StateNotifierStateProvider);
 
     expect(find.text('0'), findsNothing);
@@ -373,7 +373,7 @@ void main() {
         .firstState<ProviderScopeState>(find.byType(ProviderScope))
         .container;
 
-    final state = container.debugProviderStates
+    final state = container.debugProviderElements
         .firstWhere((s) => s.provider == provider.state);
 
     expect(state.hasListeners, true);